Refactor manager package to device
