+++ device/doc.go
@@ -1,0 +1,6 @@
+/*
+Package device implements device managers.
+
+Device managers handles the communication with devices (like the Launch.)
+*/
+package device
